% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anmData.R
\name{anmData}
\alias{anmData}
\title{anmData}
\usage{
anmData(code)
}
\arguments{
\item{code}{A character vector with the series code.}
}
\value{
A data frame containing the requested data.
}
\description{
This function fetches data from the ANM open data portal.
}
