% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_anm_data.R
\name{fetch_anm_data}
\alias{fetch_anm_data}
\title{Fetch ANM Data}
\usage{
fetch_anm_data(tema, endpoint)
}
\arguments{
\item{endpoint}{The API endpoint to fetch data from.}
}
\value{
A data frame containing the requested data.
}
\description{
This function fetches data from the ANM open data portal.
}
