% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anmData.R
\name{anmData}
\alias{anmData}
\title{Fetch ANM Data}
\usage{
anmData(tema, endpoint)
}
\arguments{
\item{endpoint}{The API endpoint to fetch data from.}
}
\value{
A data frame containing the requested data.
}
\description{
This function fetches data from the ANM open data portal.
}
